쿼리문 수정
