+++ src/main/java/com/hta2405/unite/dto/ProjectDetailDTO.java
@@ -18,8 +18,8 @@ public class ProjectDetailDTO {
     private String memberDesign; //진행률 - 업무
     private double memberProgressRate;  //진행률 - 진행률
     private boolean isManager = false;
-
     public void setIsManager(boolean isManager) {
         this.isManager = isManager;
     }
+    private String jobName; //직책
 }
