+++ src/main/java/com/hta2405/unite/service/ProjectServiceImpl.java
@@ -209,4 +209,5 @@ public class ProjectServiceImpl implements ProjectService {
 
         return colorInfo;
     }
+
 }
