+++ src/main/resources/templates/project/project_detail.html
@@ -7,6 +7,7 @@
 			<title>Insert title here</title>
 			<link rel="stylesheet" type="text/css" th:href="@{/css/projectDetail.css}">
 			<script th:src="@{/js/project_detail.js}"></script>
+			<script th:src="@{/js/project_color.js}"></script>
 			<script>
 				var contextPath = "${pageContext.request.contextPath}";
 				var projectName = "${left}";
@@ -19,8 +20,8 @@
 			<input type="hidden" th:value="${projectId}" id="project-id">
 			<input type="hidden" th:value="${memberId}" id="user-id">
 			<div class="container" style="margin-top: 50px;">
-				<div id="background" class="d-flex justify-content-between align-items-center">
-					<h3 id="text"><span th:text="${left}"></span> - <span th:text="${projectContent}" style="font-size: 20px;"></span></h3>
+				<div class="d-flex justify-content-between align-items-center background">
+					<h3 class="text"><span th:text="${left}"></span> - <span th:text="${projectContent}" style="font-size: 20px;"></span></h3>
 					<span>
 						<img th:src="@{/image/help.png}" style="width: 50px; height: 50px; cursor: pointer;" onclick="openHelpModal()">
 					</span>
@@ -61,16 +62,17 @@
 								<tbody>
 									<tr th:each="project : ${project}">
 										<input type="hidden" th:value="${project.projectId}" id="projectId">
-										<td th:text="${project.participantNames}"
+										<td th:text="|${project.participantNames} ${project.jobName}|"
 											th:data-id="${project.projectId}"
 											th:data-memberid="${project.memberId}"
-											class="clickable-participant"></td> <!-- 클릭 이벤트 추가 -->
+											class="clickable-participant">
+										</td>
 										<td>
 											<span th:classappend="${project.isManager} ? 'task-content clickable' : 'task-content'"
 												  th:data-id="${project.projectId}"
 												  th:data-memberid="${project.memberId}"
 												  th:data-content="${project.memberDesign}">
-												<span th:text="${project.memberDesign == null ? '업무를 지정해주세요' : project.memberDesign}"></span>
+												<span th:text="${project.memberDesign == null ? '업무를 지정해주세요' : project.memberDesign}">sdf</span>
 											</span>
 															</td>
 															<td>
@@ -125,7 +127,7 @@
 								<tbody>
 									<tr th:each="project : ${project2}" class="clickable-row" th:data-member-id="${project.memberId}">
 										<td>
-											<span th:text="${project.taskWriter}"></span>
+											<span th:text="|${project.taskWriter} ${project.jobName}|"></span>
 										</td>
 										<td>
 											<span th:text="${project.taskSubject}"></span>
